refactor: Simplify tests to a simple assert
Given we're using experimental module support of Node, I had to rename
the files to use the mjs extension.
