+++ package.json
@@ -2,35 +2,27 @@
   "name": "oceanwind",
   "version": "0.2.2",
   "description": "compiles tailwind like shorthand syntax into css at runtime",
-  "module": "index.js",
-  "main": "index.js",
+  "module": "index.mjs",
+  "main": "index.mjs",
   "scripts": {
     "start": "servor --browse --reload --editor",
     "build": "esbuild --bundle index.js --outfile=index.min.js --format=esm --minify",
     "prepublish": "npm run build",
-    "test": "jest"
+    "test": "node --experimental-modules test/ssr.test.mjs"
   },
   "keywords": [],
   "author": "Luke Jackson (@lukejacksonn)",
   "license": "MIT",
   "devDependencies": {
-    "@babel/core": "^7.11.6",
-    "@babel/preset-env": "^7.11.5",
-    "babel-jest": "^26.3.0",
     "esbuild": "0.6.28",
     "htm": "^3.0.4",
-    "jest": "^26.4.2",
     "preact": "^10.4.8",
     "preact-render-to-string": "^5.1.10",
     "prettier": "^2.1.1",
-    "servor": "4.0.2"
+    "servor": "4.0.2",
+    "tape": "^5.0.1"
   },
   "prettier": {},
-  "jest": {
-    "transform": {
-      "^.+\\.[t|j]sx?$": "babel-jest"
-    }
-  },
   "dependencies": {
     "otion": "^0.6.1"
   }
